@@ -46,30 +46,32 @@
     "!*/__tests__"
   ],
   "dependencies": {
-    "@loopback/authentication": "^3.2.3",
-    "@loopback/boot": "^1.5.12",
-    "@loopback/context": "^1.24.0",
-    "@loopback/core": "^1.11.0",
-    "@loopback/openapi-v3": "^1.10.2",
-    "@loopback/repository": "^1.15.5",
-    "@loopback/rest": "^1.24.0",
-    "@loopback/rest-explorer": "^1.4.5",
-    "@loopback/security": "^0.1.8",
-    "@loopback/service-proxy": "^1.3.12",
-    "loopback-connector-postgresql": "^3.8.0"
+    "@loopback/authentication": "^3.3.3",
+    "@loopback/boot": "^1.7.4",
+    "@loopback/context": "^2.1.1",
+    "@loopback/core": "^1.12.4",
+    "@loopback/openapi-v3": "^2.0.0",
+    "@loopback/repository": "^1.19.1",
+    "@loopback/rest": "^2.0.0",
+    "@loopback/rest-explorer": "^1.4.10",
+    "@loopback/security": "^0.1.13",
+    "@loopback/service-proxy": "^1.3.17",
+    "bcryptjs": "^2.4.3",
+    "loopback-connector-postgresql": "^3.8.3"
   },
   "devDependencies": {
-    "@loopback/build": "^2.1.0",
-    "source-map-support": "^0.5.16",
-    "@loopback/testlab": "^1.9.5",
-    "@types/node": "^10.17.5",
-    "@typescript-eslint/parser": "^2.9.0",
-    "@typescript-eslint/eslint-plugin": "^2.9.0",
-    "@loopback/eslint-config": "^4.1.5",
-    "eslint": "^6.7.1",
-    "eslint-config-prettier": "^6.7.0",
-    "eslint-plugin-eslint-plugin": "^2.1.0",
+    "@loopback/build": "^3.1.1",
+    "@loopback/eslint-config": "^5.0.3",
+    "@loopback/testlab": "^1.10.3",
+    "@types/bcryptjs": "^2.4.2",
+    "@types/node": "^10.17.14",
+    "@typescript-eslint/eslint-plugin": "^2.19.0",
+    "@typescript-eslint/parser": "^2.19.0",
+    "eslint": "^6.8.0",
+    "eslint-config-prettier": "^6.10.0",
+    "eslint-plugin-eslint-plugin": "^2.2.1",
     "eslint-plugin-mocha": "^6.2.2",
-    "typescript": "~3.7.2"
+    "source-map-support": "^0.5.16",
+    "typescript": "~3.7.5"
   }
 }
